+++ tests/unit/test_solvers/test_processed_variable.py
@@ -37,8 +37,8 @@ class TestProcessedVariable(unittest.TestCase):
         y_sol = np.array([np.linspace(0, 5)])
         var_casadi = to_casadi(var, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var,
-            var_casadi,
+            [var],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -51,8 +51,8 @@ class TestProcessedVariable(unittest.TestCase):
         y_sol = np.array([1])[:, np.newaxis]
         var_casadi = to_casadi(var, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var,
-            var_casadi,
+            [var],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -75,8 +75,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -84,8 +84,8 @@ class TestProcessedVariable(unittest.TestCase):
         np.testing.assert_array_equal(processed_var(t_sol, x_sol), y_sol)
         eqn_casadi = to_casadi(eqn_sol, y_sol)
         processed_eqn = pybamm.ProcessedVariable(
-            eqn_sol,
-            eqn_casadi,
+            [eqn_sol],
+            [eqn_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -104,8 +104,8 @@ class TestProcessedVariable(unittest.TestCase):
         x_s_edge.mesh = disc.mesh["separator"]
         x_s_casadi = to_casadi(x_s_edge, y_sol)
         processed_x_s_edge = pybamm.ProcessedVariable(
-            x_s_edge,
-            x_s_casadi,
+            [x_s_edge],
+            [x_s_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -120,8 +120,8 @@ class TestProcessedVariable(unittest.TestCase):
         y_sol = np.ones_like(x_sol)[:, np.newaxis]
         eqn_casadi = to_casadi(eqn_sol, y_sol)
         processed_eqn2 = pybamm.ProcessedVariable(
-            eqn_sol,
-            eqn_casadi,
+            [eqn_sol],
+            [eqn_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -155,7 +155,7 @@ class TestProcessedVariable(unittest.TestCase):
         c = pybamm.StateVector(slice(0, var_pts[x]), domain=["SEI layer"])
         c.mesh = mesh["SEI layer"]
         c_casadi = to_casadi(c, y_sol)
-        pybamm.ProcessedVariable(c, c_casadi, solution, warn=False)
+        pybamm.ProcessedVariable([c], [c_casadi], solution, warn=False)
 
     def test_processed_variable_2D_x_r(self):
         var = pybamm.Variable(
@@ -182,8 +182,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -217,8 +217,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -237,8 +237,8 @@ class TestProcessedVariable(unittest.TestCase):
         x_s_edge.secondary_mesh = disc.mesh["current collector"]
         x_s_casadi = to_casadi(x_s_edge, y_sol)
         processed_x_s_edge = pybamm.ProcessedVariable(
-            x_s_edge,
-            x_s_casadi,
+            [x_s_edge],
+            [x_s_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -271,8 +271,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -295,8 +295,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, u_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, u_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -318,8 +318,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, u_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, u_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -340,8 +340,8 @@ class TestProcessedVariable(unittest.TestCase):
         y_sol = np.array([np.linspace(0, 5, 1000)])
         var_casadi = to_casadi(var, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var,
-            var_casadi,
+            [var],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -353,8 +353,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         eqn_casadi = to_casadi(eqn, y_sol)
         processed_eqn = pybamm.ProcessedVariable(
-            eqn,
-            eqn_casadi,
+            [eqn],
+            [eqn_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -377,8 +377,8 @@ class TestProcessedVariable(unittest.TestCase):
         y_sol = np.array([[100]])
         eqn_casadi = to_casadi(eqn, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            eqn,
-            eqn_casadi,
+            [eqn],
+            [eqn_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -401,8 +401,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -421,8 +421,8 @@ class TestProcessedVariable(unittest.TestCase):
         )
         eqn_casadi = to_casadi(eqn_sol, y_sol)
         processed_eqn = pybamm.ProcessedVariable(
-            eqn_sol,
-            eqn_casadi,
+            [eqn_sol],
+            [eqn_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -441,8 +441,8 @@ class TestProcessedVariable(unittest.TestCase):
         x_casadi = to_casadi(x_disc, y_sol)
 
         processed_x = pybamm.ProcessedVariable(
-            x_disc,
-            x_casadi,
+            [x_disc],
+            [x_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -455,8 +455,8 @@ class TestProcessedVariable(unittest.TestCase):
         r_n.mesh = disc.mesh["negative particle"]
         r_n_casadi = to_casadi(r_n, y_sol)
         processed_r_n = pybamm.ProcessedVariable(
-            r_n,
-            r_n_casadi,
+            [r_n],
+            [r_n_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -479,8 +479,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         eqn_casadi = to_casadi(eqn_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            eqn_sol,
-            eqn_casadi,
+            [eqn_sol],
+            [eqn_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -517,8 +517,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -565,8 +565,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -600,8 +600,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -629,8 +629,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -668,8 +668,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -692,8 +692,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, u_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, u_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -736,8 +736,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, u_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, u_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -813,8 +813,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -836,8 +836,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         var_casadi = to_casadi(var_sol, y_sol)
         processed_var = pybamm.ProcessedVariable(
-            var_sol,
-            var_casadi,
+            [var_sol],
+            [var_casadi],
             pybamm.Solution(t_sol, y_sol, pybamm.BaseModel(), {}),
             warn=False,
         )
@@ -866,8 +866,8 @@ class TestProcessedVariable(unittest.TestCase):
 
         with self.assertRaisesRegex(NotImplementedError, "Shape not recognized"):
             pybamm.ProcessedVariable(
-                var_sol,
-                var_casadi,
+                [var_sol],
+                [var_casadi],
                 pybamm.Solution(t_sol, u_sol, pybamm.BaseModel(), {}),
                 warn=False,
             )